Layout_TestCase/http/tests/xmlhttprequest/simple-cross-origin-denied-events-post-sync-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/reply.xml. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
Test that a simple cross-origin request to a server that responds (but does not permit cross-origin requests) is indistinguishable from one that does not exist. Should say PASS:

PASS
